refactor: even more refactor
