--- src/Config/EventsConfig/EventsKillConfig.cs
@@ -1,16 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-public class EventsKillConfig
-{
-    [JsonPropertyName("HP")]
-    public int Hp { get; set; } = 2;
-
-    [JsonPropertyName("HeadshotHP")]
-    public int HeadshotHp { get; set; } = 3;
-
-    [JsonPropertyName("Money")]
-    public int Money { get; set; } = 200;
-
-    [JsonPropertyName("HeadshotMoney")]
-    public int HeadshotMoney { get; set; } = 300;
-}
