--- src/Config/Messages/MessagesConfig.cs
@@ -1,11 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-public class MessagesConfig
-{
-    [JsonPropertyName("ConnectChat")]
-    public string ConnectChat { get; set; } = "VIP {playername} joined the server";
-
-    [JsonPropertyName("DisconnectChat")]
-    public string DisconnectChat { get; set; } = "VIP {playername} left the server";
-
-}
