+++ src/Config/VipGroupConfig/MessagesConfig/MessagesConfig.cs
@@ -1,0 +1,13 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Core.Config
+{
+    public class MessagesConfig
+    {
+        [JsonPropertyName("ConnectChat")]
+        public string ConnectChat { get; set; } = "VIP {playername} joined the server";
+
+        [JsonPropertyName("DisconnectChat")]
+        public string DisconnectChat { get; set; } = "VIP {playername} left the server";
+    }
+}
