+++ src/Models/SettingsData.cs
@@ -1,9 +1,12 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Core.Models;
-
-public class SettingsData
+namespace Core.Models
 {
-    [JsonPropertyName("Prefix")]
-    public string Prefix { get; set; } = "[VIP-Plugin] ";
+    public class SettingsData
+    {
+        [JsonPropertyName("Prefix")]
+        public string Prefix { get; set; } = "[VIP-Plugin] ";
+    }
+
 }
+
