+++ src/VIP-Plugin.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.9.34701.34
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VIP-Plugin", "VIP-Plugin.csproj", "{872E2F62-26F9-4372-809B-C4F68FFAB730}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VIP-Plugin", "VIP-Plugin.csproj", "{7347465E-B70E-4A1D-96FB-BD32058F4CD5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{872E2F62-26F9-4372-809B-C4F68FFAB730}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{872E2F62-26F9-4372-809B-C4F68FFAB730}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{872E2F62-26F9-4372-809B-C4F68FFAB730}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{872E2F62-26F9-4372-809B-C4F68FFAB730}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7347465E-B70E-4A1D-96FB-BD32058F4CD5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7347465E-B70E-4A1D-96FB-BD32058F4CD5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7347465E-B70E-4A1D-96FB-BD32058F4CD5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7347465E-B70E-4A1D-96FB-BD32058F4CD5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {19942C4B-74D8-4D08-B9D4-189C65A3C52F}
+		SolutionGuid = {FEAAD9E4-116D-45D1-86C8-018BCCFBFC6A}
 	EndGlobalSection
 EndGlobal
